src/models/database.py
import asyncpg
import json
from typing import Dict, Any, List, Optional
from datetime import datetime
import uuid
from ..utils.config import config

class DatabaseManager:
    def __init__(self):
        self.pool = None
        
    async def connect(self):
        """Create connection pool"""
        # TODO: Implement database connection
        pass
        
    async def create_task(self, product_name: str, sources: List[str] = None) -> Dict[str, Any]:
        """Create a new analysis task"""
        # TODO: Implement task creation
        pass
    
    # TODO: Add more database methods

# Global database instance
db = DatabaseManager()
