requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
praw==7.7.1
google-api-python-client==2.108.0
transformers==4.35.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
beautifulsoup4==4.12.2
